pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "erasmus-workspace"
version = "0.1.0"
description = "AI Context Watcher for Development"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pysnooper>=1.2.2",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "watchdog>=4.0.2",
    "openai>=1.74.0",
    "mcp[cli]>=1.6.0",
]

[project.scripts]
erasmus = "erasmus.__main__:main"

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "hypothesis>=6.95.0",
]
dev = [
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["erasmus"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-v --cov=src"
testpaths = [
    "tests",
]
python_files = ["test_*.py"]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["erasmus"]
ignore = ["G004", "TRY003", "TRY002"]

# Enable Pyflakes (`F`), isort (`I`), and many other linters
select = ["E", "F", "I", "W", "N", "B", "A", "COM", "C4", "DTZ", "T10", "ISC", "ICN", "G", "PIE", "PYI", "PT", "RET", "SIM", "TID", "ARG", "PTH", "ERA", "PD", "PL", "TRY", "UP", "RUF"]

# Allow autofix for all enabled rules (when `--fix`) is provided
fixable = ["ALL"]
unfixable = []

# Exclude a variety of commonly ignored directories
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"] # Ignore unused imports in __init__.py
"tests/**/*.py" = ["S101"] # Allow assert in tests

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.isort]
known-first-party = ["erasmus"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
show_error_codes = true
pretty = true
follow_imports = "silent"
strict_optional = true
warn_redundant_casts = true
warn_unreachable = true
warn_no_return = true

# Ignore errors in certain files
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
no_implicit_optional = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.uv.workspace]
members = [
    "erasmus",
    "packager"
    ]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.24.0",
]
